Crate proxer

proxer-rs (proxer-api <-> proxer-rs <-> your_rust_application)

Inoffizielle Bibliothek zur proxer-api in rust.

Die meisten Dokumentationen sind aus dem Proxer-API wiki. Sollte etwas in der crate fehlen, oder anderweitig nicht stimmen, bitte auf Github melden.

Aktuell nur auf nighly verfügbar.

Auszug aus dem Proxer-API wiki

Die Proxer API ist eine Programmierschnittstelle, die dazu dient, Proxer in allen Aspekten erweiterbar zu machen.

Nutzungsbedingungen

Die Anwendung muss ausdrücklich als "inoffizielle Anwendung" deklariert werden (z.B. beim Starten der Anwendung in Form eines Dialogs, in der Anwendungsbeschreibung, im Impressum und/oder in den Credits).

Die Anzahl der Anfragen an den Proxer-Server muss so gering gehalten werden wie nur möglich. *Auf die Sicherheit der Anwendung muss stark geachtet werden. Falls möglich, soll immer SSL zur Verschlüsselung der Kommunikation verwendet werden. Gemeldete Sicherheitsprobleme müssen in kürzester Zeit behoben werden.

In besonderen Fällen kann genesis zusätzliche Forderungen stellen, die dann ebenfalls eingehalten werden müssen.

Die Nutzungsbedingungen könnten sich jederzeit ändern. Nach einer Änderung müssen die Änderungen innerhalb von 30 Tagen umgesetzt werden. Alle Änderungen werden in der Mailing-List erläutert.

Finanzielles

Nehmen wir an, es gibt eine "Kodi" Anwendung. Wenn alle diese Anwendung verwenden würden, könnten wir unsere Server nicht mehr bezahlen. Das wäre nachteilhaft für alle beteiligten. Programmiere deine Anwendungen so, dass du es mit bestem Gewissen verwenden und verteilen kannst. Prinzipiell werden wir dich nicht davon abhalten, z.B. mit Hilfe von Werbung Geld mit deiner Anwendung zu verdienen. Unsere Bitte an dieser Stelle ist, dass ihr unsere Spenden-Adresse irgendwo in die Anwendung einbaut (bei Android Anwendungen z.B. in das Menü), oder ein Teil der Einnahmen spendet, damit wir weiterhin imstande sind unsere Server zu finanzieren.

Verwaltung

Mailing-List Der Eintrag in der Mailing-List ist freiwillig. Du kannst dich in unsere Mailing-List eintragen lassen um per Proxer-PN über eine Änderung an den Nutzungsbedingungen oder der API benachrichtigt zu werden. Schreibe hierfür eine kurze PN an genesis mit dem Inhalt: "Ich möchte in der API-Mailingliste eingetragen werden". Wir laden dich dann in eine Proxer-Gruppe ein.

Open Source Open-Source Projekte haben die Möglichkeit, "mehr" als nur eine inoffizielle Anwendung zu bleiben. Falls das Projekt z.B. auf github oder einer anderen Versionsverwaltung gehostet ist, besteht die Möglichkeit, dass wir die Anwendung zu unseren offiziellen Anwendungen aufnehmen. Offizielle Anwendungen werden wir beispielsweise auf unserer Facebook-Seite empfehlen und aktiv unterstützen.

Support Proxer stellt für den Support deiner Anwendung eine Forum-Kategorie zur Verfügung. Hier kannst du für deine Anwendung einen Thread erstellen, welcher für Fragen, Vorschläge, Probleme und Kritik offen steht. In den ersten Beitrag beschreibst du deine Anwendung, postest dein Changelog, wie man deine Anwendung benutzt und ggf. weitere Links wie z.B. zum git-repository. Werbung für deine private Webseite sind strengstens untersagt und ist gegen die Forumsregeln. Für sonstige Diskussionen und initiierung neuer Projekte haben wir ebenfalls eine eigene Kategorie.

Datenschutz

Passwörter Eingegebene Passwörter sollten stets entweder in geschützten Speicherbereichen (z.B. SharedPreferences in Android) und/oder verschlüsselt gespeichert werden.

Statistiken Du darfst gerne Google Analytics verwenden, wir empfehlen jedoch die datenschutzfreundliche(https://www.datenschutzbeauftragter-info.de/fachbeitraege/google-analytics-datenschutzkonform-einsetzen/) Version zu verwenden.

Transparenz Deine Anwendung sollte möglichst transparent arbeiten und keine dubiosen Aktionen im Hintergrund durchführen. Wir werden in bestimmten Abständen Anwendungen prüfen und bei festgestelltem Missbrauch der Schnittstelle die Erlaubnis zur Nutzung der Anwendung entziehen.

Technische Informationen

Ressourcen Wir bitten ausdrücklich, sparsam mit unseren Ressourcen umzugehen. Das ist sehr wichtig, da unsere Server besonders an Wochenenden an ihren Grenzen sind. Um dies zu erzielen, muss darauf geachtet werden, dass keine Anfrage unnötig gesendet wird. Im Falle einer Anwendung, die darstellt ob eine neue Episode eines Animes erschienen ist, würde es beispielsweise ausreichen, wenn die Abfrage höchstens alle 30 Minuten durchgeführt wird.

Firewall Proxer verwendet Cloudflare, um im Falle eines Ddos-Angriffes die Aufrechterhaltung unserer Dienste zu gewährleisten. Aktuell besteht leider keine Möglichkeit, diese zu umgehen. Fehlerhafte Anfragen sollten daher abgefangen werden, damit die Anwendung im Falle eines Angriffes nicht abstürzt.

Dokumentation der Schnittstellen

Im Juni 2016 wurde eine Neuentwicklung der Schnittstellen durchgeführt. Die alten Schnittstellen werden ab sofort nicht mehr unterstüzt, bleiben jedoch bis Januar 2017 weiterhin aktiv. Die neue API ermöglicht eine Versionisierung der Schnittstellen, sodass eine nahtlose Weiterentwicklung stattfindet.

Proxer-API Version 1: Unterstützung ab Juni 2016. Aktuelle Version.

Weitere Fragen?

Du kannst dich jederzeit an genesis(http://proxer.me/wiki/Benutzer:Genesis) wenden.

(Quelle:http://proxer.me/wiki/Proxer_API, vom 20.01.2017 18:00)

Modules

anime

error

info

list

manga

media

messenger

models

notification

ucp

user

Macros

check_data

check_error

param_build

url

Structs

OldNews
Ermöglicht den Zugriff auf die Proxer News. Hierbei werden pro Seite höchstens 15 News ausgegeben.

OldNotification
Beinhaltet eine einzelene News bzw. Notification (Benutzt die alte api Schnittstelle?).

Proxer
Klasse um die http/s Verbindungen aufzubauen und die API-Daten zu verwalten.