Struct proxer::anime::Anime

pub struct Anime<'anime> { /* fields omitted */ }

Diese Klasse dient dazu, Streams für Animes zu erhalten, und andere rein Anime-bezogene Aktionen durchzuführen. Der Zugriff auf diese Klasse ist stark beschränkt.

Methods

impl<'anime> Anime<'anime>

fn get_streams(&self,
p_id: u64,
p_episode: u64,
p_language: Language)
-> Result<Vec<Stream>>

Diese Funktion liefert die Streams einer Folge eines festgelegten Animes, ohne den Proxerstream!. Um Animepunkte für das Schauen zu erhalten muss ein User angemeldet sein.

Arguments

p_id - Die id des Entrys

p_episode - Die Episodennummer der zu ladenden Folge

p_language - Die zu ladende Sprache (gersub,gerdub,engsub,engdub)

fn get_proxerstreams(&self,
p_id: u64,
p_episode: u64,
p_language: Language)
-> Result<Vec<ProxerStream>>

Diese Funktion liefert die Streams einer Folge eines festgelegten Animes, inklusive des Proxerstreams!. Um Animepunkte für das Schauen zu erhalten muss ein User angemeldet sein. Die Schnittstelle liefert auch alle weiteren Streams, eine zusätzliche Abfrage der "Get Streams" Schnittstelle ist also nicht nötig. Hinweis: Diese Schnittstelle ist momentan gesperrt!

Arguments

p_id - Die id des Entrys

p_episode - Die Episodennummer der zu ladenden Folge

p_language - Die zu ladende Sprache (gersub,gerdub,engsub,engdub)

fn get_link(&self, p_id: u64) -> Result<String>

Diese Funktion dient dazu, den Link zu einem einzelnen Stream zu erhalten. Dazu muss die ID des gewählten Streams angegeben werden. Dieser kann erhalten werden durch die Get Streams oder Get Proxerstreams Funktionen. Diese Funktion erhöht euch den Viewcount des Streams.

Arguments

p_id - Die id des Entrys