Module proxer::info

Structs

Comment
Diese Funktion liefert die für einen Anime/Manga abgegebenen Kommentare (mit mehr als 300 Zeichen).

Entry
Diese Funktion liefert die Daten eines Animes/Mangas anhand seiner ID (Dies bezieht sich NUR auf die Kerndaten des Anime, nicht die Daten, die über die folgenden Funktionen abgefragt werden können)

EntryTag
Diese Funktion liefert alle Tags eines Entrys.

FullEntry

FullEntryDataGroups

FullEntryDataPublisher

FullEntryDataSeasons

FullEntryDataTags

Group
Diese Funktion liefert die für einen Anime/Manga Eingetragenen Übersetzergruppen.

Industry
Diese Funktion liefert alle Daten zu einer Firma anhand ihrer ID.

Info
Diese Klasse beinhaltet alle Schnittstellen, die sich auf das Info-System Proxers beziehen, insbesondere in Bezug auf Informationen zu Animes und Mangas, aber auch zu den weiteren Verzeichnissen.

ListInfo
Diese Funktion liefert eine Liste aller Episoden/Kapitel eines Entrys anhand dessen ID.

ListInfoDataEpisode

Name
Diese Funktion liefert die unterschiedlichen Synonyme eines Animes/Mangas anhand seiner ID

Publisher
Diese Funktion liefert die für einen Anime/Manga Eingetragenen Publisher.

Relation
Diese Funktion liefert alle Verbindungen eines Entrys.

Season
Diese Funktion liefert die für einen Anime/Manga Eingetragenen Seasons.

TranslatorGroup
Diese Funktion liefert alle Daten zu einer Sub/Scanlation Gruppe anhand ihrer ID.