Struct proxer::info::Info

pub struct Info<'info> { /* fields omitted */ }

Diese Klasse beinhaltet alle Schnittstellen, die sich auf das Info-System Proxers beziehen, insbesondere in Bezug auf Informationen zu Animes und Mangas, aber auch zu den weiteren Verzeichnissen.

Methods

impl<'info> Info<'info>

fn get_fullentry(&self, p_id: u64) -> Result<FullEntry>

Diese Funktion liefert ALLE Daten eines Animes/Mangas anhand seiner ID. Diese Schnittstelle belastet den Server stark und sollte nur verwendet werden, wenn auch wirklich ALLE Informationen des Animes benötigt werden (wie etwa bei Detailseiten).

Arguments

id - Die ID des gewünschten Animes/Mangas.

fn get_entry(&self, p_id: u64) -> Result<Entry>

Diese Funktion liefert die Daten eines Animes/Mangas anhand seiner ID (Dies bezieht sich NUR auf die Kerndaten des Anime, nicht die Daten, die über die folgenden Funktionen abgefragt werden können)

Arguments

id - Die ID des gewünschten Animes/Mangas.

fn get_name(&self, p_id: u64) -> Result<Vec<Name>>

Diese Funktion liefert die unterschiedlichen Synonyme eines Animes/Mangas anhand seiner ID

Arguments

id - Die ID des gewünschten Animes/Mangas.

fn get_gate(&self, p_id: u64) -> Result<bool>

Diese Funktion liefert einen boolean Wert zurück, ob der per ID spezifizierte Anime/Manga eine ab-18 Prüfung erfordert.

Arguments

id - Die ID des gewünschten Animes/Mangas.

fn get_language(&self, p_id: u64) -> Result<Vec<String>>

Diese Funktion liefert die für einen Anime/Manga Eingetragenen Sprachen (in denen der Eintrag auf Proxer erhältlich ist).

Arguments

id - Die ID des gewünschten Animes/Mangas.

fn get_season(&self, p_id: u64) -> Result<Vec<Season>>

Diese Funktion liefert die für einen Anime/Manga Eingetragenen Seasons.

Arguments

id - Die ID des gewünschten Animes/Mangas.

fn get_groups(&self, p_id: u64) -> Result<Vec<Group>>

Diese Funktion liefert die für einen Anime/Manga Eingetragenen Übersetzergruppen.

Arguments

id - Die ID des gewünschten Animes/Mangas.

fn get_publisher(&self, p_id: u64) -> Result<Vec<Publisher>>

Diese Funktion liefert die für einen Anime/Manga Eingetragenen Publisher.

Arguments

id - Die ID des gewünschten Animes/Mangas.

fn get_listinfo(&self,
p_id: u64,
p_page: Option<u64>,
p_limit: Option<u64>)
-> Result<ListInfo>

Diese Funktion liefert eine Liste aller Episoden/Kapitel eines Entrys anhand dessen ID.

Arguments

id - Die ID des gewünschten Animes/Mangas.

p - (optional): Die zu ladende Seite der Liste, Start bei 0. Default 0.

limit - (optional): Die Nummer der zu ladenden Episoden/Kapitel pro Seite. Default 50.

fn get_comments(&self,
p_id: u64,
p_page: Option<u64>,
p_limit: Option<u64>,
p_sort: Option<Sort>)
-> Result<Vec<Group>>

Diese Funktion liefert die für einen Anime/Manga abgegebenen Kommentare (mit mehr als 300 Zeichen).

Arguments

id - Die ID des gewünschten Animes/Mangas.

p - Die zu ladende Seite der Liste, Start bei 0. Default 0.

limit - Die Nummer der zu ladenden Kommentare pro Seite. Default 25.

sort - Ändert die Sortierung der Liste. Setze Wert "rating" um nach meisten Empfehlungen zu sortieren, ansonsten Sortierung nach Neueste zuerst.

fn get_relations(&self, p_id: u64) -> Result<Vec<Relation>>

Diese Funktion liefert alle Verbindungen eines Entrys.

Arguments

id - Die ID des gewünschten Animes/Mangas.

fn get_entrytag(&self, p_id: u64) -> Result<Vec<EntryTag>>

Diese Funktion liefert alle Tags eines Entrys.

Arguments

id - Die ID des gewünschten Animes/Mangas.

fn get_translatorgroup(&self, p_id: u64) -> Result<TranslatorGroup>

Diese Funktion liefert alle Daten zu einer Sub/Scanlation Gruppe anhand ihrer ID.

Arguments

id - Die ID der gewünschten Gruppe.

fn get_industry(&self, p_id: u64) -> Result<Industry>

Diese Funktion liefert alle Daten zu einer Firma anhand ihrer ID.

Arguments

id - Die ID der gewünschten Gruppe.

fn set_userinfo(&self, p_id: u64, p_type: WatchType) -> Result<()>

Diese Funktion setzt einen per ID spezifizierten Anime/Manga auf eine der Listen des Users, abhängig vom Paramter "type".

Arguments

id - Die ID des gewünschten Animes/Mangas.

type - Die Liste, zu der der Anime hinzugefügt werden soll. Erlaubt: "note" (Wird noch geschaut), "favor" (Favoriten), "finish" (Abgeschlossen)