Module proxer::list

Structs

EntryList
Diese Funktion liefert eine Liste aller Entrys einer Kategorie mit bestimmten Einschränkungsmöglichkeiten.

EntrySearch
Diese Funktion erfüllt die Aufgabe der erweiterten Suche

Industry
Diese Funktion liefert eine Liste aller Firmen, anhand bestimmter Kriterien.

IndustrysProject
Diese Funktion liefert eine Liste aller Projekte (=Entrys) einer Firma anhand ihrer ID.

List
Diese Klasse dient als Hauptmethode um die Daten von Entrys zu erhalten, insbesondere der ID (die für jede spezielle Entry Schnittstelle nötig ist). Sie enthält zwei Arten von Schnittstellen: Eine "Search" Schnittstelle, die wie die Erweiterte Suche auf Proxer funktioniert, sowie Auflistungs-Schnittstellen, die Animelisten nach bestimmten Schemata erstellen.

Tag
Diese Funktion liefert eine Liste aller Tags, anhand bestimmter Kriterien.

TagIDs
Diese Funktion zieht aus einem String die IDs aller darin vorkommenden Tags und gibt sie zurück.

TranslatorGroup
Diese Funktion liefert eine Liste aller Sub/Scanlation Gruppen, anhand bestimmter Kriterien.

TranslatorGroupProject
Diese Funktion liefert eine Liste aller Projekte (=Entrys) einer Gruppe anhand ihrer ID.