Struct proxer::list::List

pub struct List<'list> { /* fields omitted */ }

Diese Klasse dient als Hauptmethode um die Daten von Entrys zu erhalten, insbesondere der ID (die für jede spezielle Entry Schnittstelle nötig ist). Sie enthält zwei Arten von Schnittstellen: Eine "Search" Schnittstelle, die wie die Erweiterte Suche auf Proxer funktioniert, sowie Auflistungs-Schnittstellen, die Animelisten nach bestimmten Schemata erstellen.

Methods

impl<'list> List<'list>

fn entry_search(&self,
p_name: Option<String>,
p_language: Option<String>,
p_type: Option<Medium>,
p_genre: Option<String>,
p_nogenre: Option<String>,
p_fsk: Option<String>,
p_sort: Option<SearchSort>,
p_length: Option<u64>,
p_lengthlimit: Option<LengthLimit>,
p_tags: Option<String>,
p_notags: Option<String>,
p_tagratefilter: Option<String>,
p_tagspoilerfilter: Option<String>,
p_page: Option<u64>,
p_limit: Option<u64>)
-> Result<Vec<EntrySearch>>

Diese Funktion erfüllt die Aufgabe der erweiterten Suche

Arguments

p_name - Ein zu suchender Entryname. Die Schnittstelle sucht sowohl nach exakten Matches als auch nach Teil-Matches im Titel des Entrys (oder einem seiner Synonyme)

p_language - Die zu suchende Sprache. Erlaubte Werte: "de","en". Default: Beide.

p_type - Der zu suchende Typ. Erlaubte Werte: 'p_animeseries', 'movie', 'ova', 'mangaseries', 'oneshot', 'doujin', 'hentai', 'hmanga', 'p_all-anime'(kein H), 'all-manga'(kein H), 'all' (Default, kein H), 'all18'(H)

p_genre - Genre, die der Entry enthalten soll. Als Leerzeichen (oder Plus "+") separierter String, Genre-Namen wie in Suche. (Also ausgeschrieben, z.B. "Action") (Kein Effekt wenn leer)

p_nogenre - Genre, die der Entry nicht enthalten darf. Angabe wie genre. (Kein Effekt wenn leer)

p_fsk - Die zu suchenden Gefahrensymbole/Altersbeschränkungen. Erlaubte Werte: "fsk0", "fsk6", "fsk12", "fsk16", "fsk18", "bad_language", "violence", "fear", "sex". Anzugeben als Leerzeichen oder Plus ("+") separierter String. (Kein Effekt wenn leer)

p_sort - Wie die Ergebnisse sortiert werden sollen. Erlaubte Werte: "relevance" (Qualität des Name-Matches, Default), "clicks", "rating" (Sortierung erst nach Anzahl und dann Wertung der Stimmen), "count" (Anzahl der Kapitel/Episoden), "name" (Alphabetisch)

p_length - Die Anzahl der Kapitel/Episoden (Müssen nicht hochgeladen sein!), die ein Entry mindestens/höchstens haben darf. Zwischen 0 und 400 (exklusive). (Kein Effekt wenn leer)

p_length-limit - Ob der Parameter "length" als Minimal- oder Maximalwert verwendet werden soll. Erlaubte Werte: "up" (größer gleich), "down" (kleiner gleich, Default). Hat keinen Effekt wenn "length" nicht korrekt gesetzt ist.

p_tags - Tags, die der Entry enthalten soll. Als Leerzeichen (oder Plus "+") separierter String von Tag-Ids. (Kein Effekt wenn leer)

p_notags - Tags, die der Entry nicht enthalten darf. Als Leerzeichen (oder Plus "+") separierter String von Tag-Ids. (Kein Effekt wenn leer)

p_tagratefilter - Welche Art Tags berücksichtigt werden soll (negativ als auch positiv), "rate_1" für nur eingetragene Tags (Default), "rate_10" für zusätzlich Unbestimmte Tags.

p_tagspoilerfilter - Inwieweit Spoiler-Tags berücksichtigt werden sollen (negativ als auch positiv), "spoiler_0" für keine Spoiler (Default), "spoiler_10" für Spoiler und Nicht-Spoiler, "spoiler_1" für nur Spoiler.

p_page - Die zu ladende Seite, Beginn bei 0, Default 0.

p_limit - Wie viele Einträge eine Seite enthalten soll. Default 100.

fn get_entrylist(&self,
p_kat: Option<Kategorie>,
p_medium: Option<Medium>,
p_is_h: Option<bool>,
p_start: Option<String>,
p_page: Option<u64>,
p_limit: Option<u64>)
-> Result<Vec<EntryList>>

Diese Funktion liefert eine Liste aller Entrys einer Kategorie mit bestimmten Einschränkungsmöglichkeiten.

Arguments

p_kat - Die Kategorie des Entrys. Erlaubte Werte: "anime","manga". Default: "anime"

p_medium - Welche Art medium geladen werden soll. Erlaubte Werte: "animeseries", "movie", "ova", "hentai", "mangaseries", "oneshot", "doujin", "hmanga". (Kein Effekt wenn leer)

p_is_h - Ob die Liste Hentais enthalten soll. "true" für Hentai-Liste, sonst non-Hentai Liste (default). Dieser Parameter hat keinen Effekt wenn "medium" gesetzt ist

p_start - Mit welchem String der Name der Entrys beginnen soll. Nützlich, um z.B. nach Anfangsbuchstaben zu filtern. Um nach nicht-alphabetischen Anfängen (Erstes Zeichen) zu filtern, 'nonAlpha' angeben. (Kein Effekt wenn leer)

p_page - Die zu ladende Seite, Beginn bei 0, Default 0.

p_limit - Wie viele Einträge eine Seite enthalten soll. Default 100.

fn get_tag_ids(&self, p_search: String) -> Result<TagIDs>

Diese Funktion zieht aus einem String die IDs aller darin vorkommenden Tags und gibt sie zurück.

Arguments

search - Ein beliebiger String. Sollte sinnvollerweise Tags enthalten, einzelne Tags durch Leerzeichen getrennt (also keine Kommata zur Trennung oder dergleichen verwenden). Es können nur Tags erkannt werden, vor und nach denen ein Leerzeichen bzw der Anfang/das Ende des Strings sind. Zudem darf vor einem Tag (Also nach dem Leerzeichen/Beginn des Strings vor dem Tag) ein Minus ("-") stehen. Solcherart gekennzeichnete Tags werden gesondert ausgegeben.

fn get_tags(&self,
p_search: Option<String>,
p_type: Option<String>,
p_sort: Option<String>,
p_sort_type: Option<String>,
p_sub_type: Option<SubType>)
-> Result<Vec<Tag>>

Diese Funktion liefert eine Liste aller Tags, anhand bestimmter Kriterien.

Arguments

p_search - Ein beliebiger String. Nur Tags, deren Name oder Beschreibung diesen String als Substring haben oder gleich ihm sind sind werden ausgegeben.

p_type - Welcher Typ Tag angezeigt werden soll. Mögliche Werte: entry_genre (Genres, momentan nicht in Benutzung), entry_tag (normale Tags), entry_tag_h (H-tags), gallery (Gallerie Tags, momentan nicht in Benutzung). Default: Alles außer H-Tags.

p_sort - Nach welchem Ausgabeparameter die Liste sortiert werden soll. Erlaubte Werte: id, type, tag, description, blacklist, subtype. Default: tag

p_sort_type - In welcher Reihenfolge die Sortierung ist. 'ASC' für aufsteigend, 'DESC' für absteigend (Jeder andere Wert wird zu DESC). Default: ASC

p_subtype - Die Kategorie des Tags

fn get_translatorgroups(&self,
p_start: Option<String>,
p_contains: Option<String>,
p_page: Option<u64>,
p_limit: Option<u64>)
-> Result<Vec<TranslatorGroup>>

Diese Funktion liefert eine Liste aller Sub/Scanlation Gruppen, anhand bestimmter Kriterien.

Arguments

p_start - Mit welchem String der Name der Gruppen beginnen soll. Nützlich, um z.B. nach Anfangsbuchstaben zu filtern. Um nach nicht-alphabetischen Anfängen (Erstes Zeichen) zu filtern, 'nonAlpha' angeben. (Kein Effekt wenn leer)

p_contains - Nur Gruppen, die diesen String als Substring ihres Namens haben, werden ausgegeben. (Kein Effekt wenn leer) country (optional): Ermöglich, Gruppen nach Sprache zu filtern, erlaubte Werte: "de", "en", "misc". Default: Alle.

p_page - Die zu ladende Seite, Beginn bei 0, Default 0.

p_limit - Wie viele Einträge eine Seite enthalten soll. Default 100.

fn get_industrys(&self,
p_start: Option<String>,
p_contains: Option<String>,
p_country: Option<String>,
p_type: Option<Firma>,
p_page: Option<u64>,
p_limit: Option<u64>)
-> Result<Vec<Industry>>

Diese Funktion liefert eine Liste aller Firmen, anhand bestimmter Kriterien.

Arguments

p_start - Mit welchem String der Name der Firma beginnen soll. Nützlich, um z.B. nach Anfangsbuchstaben zu filtern. Um nach nicht-alphabetischen Anfängen (Erstes Zeichen) zu filtern, 'nonAlpha' angeben.(Kein Effekt wenn leer)

p_contains - Nur Firmen, die diesen String als Substring ihres Namens haben, werden ausgegeben. (Kein Effekt wenn leer)

p_country - Ermöglich, Firmen nach Sprache zu filtern, erlaubte Werte: "de", "us", "jp", "misc". Default: Alle.

p_type - Ermöglicht, Firmen nach Typ zu filtern, erlaubte Werte: 'publisher','studio','producer','record_label','talent_agent','streaming'. Default: Alle.

p_page - Die zu ladende Seite, Beginn bei 0, Default 0.

p_limit - Wie viele Einträge eine Seite enthalten soll. Default 100.

fn get_translatorgroups_projects(&self,
p_id: u64,
p_type: Option<TranslationStatus>,
p_is_h: Option<i8>,
p_page: Option<u64>,
p_limit: Option<u64>)
-> Result<Vec<TranslatorGroupProject>>

Diese Funktion liefert eine Liste aller Projekte (=Entrys) einer Gruppe anhand ihrer ID.

Arguments

p_id - Die ID der gewünschten Gruppe.

p_type - Ein Integer. Wenn gesetzt werden nur Entrys des gegebenen Übersezungs-Status ausgegeben (0=undefined, 1=abgeschlossen, 2=am laufen, 3=Geplant, 4=abgebrochen, 5=lizenziert). Default: Alle.

p_is_h - Ein Integer. Steuert die Ausgabe von H-Inhalten. Werte: -1 (kein H, Default), 0 (beides), 1 (nur H)

p_page - Die zu ladende Seite, Beginn bei 0, Default 0.

p_limit - Wie viele Einträge eine Seite enthalten soll. Default 100.

fn get_industry_projects(&self,
p_id: u64,
p_type: Option<Firma>,
p_is_h: Option<i8>,
p_page: Option<u64>,
p_limit: Option<u64>)
-> Result<Vec<IndustrysProject>>

Diese Funktion liefert eine Liste aller Projekte (=Entrys) einer Firma anhand ihrer ID.

Arguments

p_id - Die ID der gewünschten Firma.

p_type - Ermöglicht, Entrys nach Typ der Firma zu filtern (eine Firma kann als verschiedene Typen auftreten), erlaubte Werte: 'publisher','studio','producer','record_label','talent_agent','streaming'. Default: Alle.

p_is_h - Ein Integer. Steuert die Ausgabe von H-Inhalten. Werte: -1 (kein H, Default), 0 (beides), 1 (nur H)

p_page - Die zu ladende Seite, Beginn bei 0, Default 0.

p_limit - Wie viele Einträge eine Seite enthalten soll. Default 100.