Struct proxer::manga::Chapter

pub struct Chapter {
    pub cid: u64,
    pub eid: u64,
    pub title: String,
    pub uploader: u64,
    pub username: String,
    pub timestamp: i64,
    pub tid: u64,
    pub tname: String,
    pub server: u32,
    pub pages: Vec<Vec<(i32, i32, i32)>>,
}

Diese Funktion liefert ein Kapitel eines festgelegten Mangas. Um Mangapunkte für das Lesen zu erhalten muss ein User angemeldet sein.

Fields
cid: u64

Die ID des Kapitels

eid: u64

Die ID des Entrys

title: String

Der Titel des Kapitels

uploader: u64

Die ID des Uploaders

username: String

Der Name des Uploaders

timestamp: i64

Der Hochladezeitpunkt (Unix-Timestamp als Integer)

tid: u64

Die ID der zugewiesenen Scangruppe, null wenn nicht vorhanden.

tname: String

Der Name der zugewiesenen Scangruppe, null wenn nicht vorhanden.

server: u32

Notwendig zur Anzeige der Seiten, siehe "pages"

pages: Vec<Vec<(i32, i32, i32)>>

Ein Array aus Arrays, die die Seiten des Kapitels in geordneter Reihenfolge enthalten. Jedes Unter-Array enthält dabei folgende Daten: 0 => Name der Seite (Dateiname), 1 => Höhe der Seite, 2 => Breite der Seite. Um eine einzelne Seite anzuzeigen, ist folgender Link nötig: //manga.proxer.me/f/// (Unter der Annahme, dass page = pages[i])

Methods

impl Chapter

fn get_chapter_link(&self) -> String