Struct proxer::messenger::Messages

pub struct Messages {
    pub message_id: u64,
    pub conference_id: u64,
    pub user_id: u64,
    pub username: String,
    pub message: String,
    pub action: String,
    pub timestamp: i64,
    pub device: String,
}

Gibt die letzten Nachrichten einer Konferenz/eines Benutzers zurück.

Fields
message_id: u64

conference_id=0 und message_id=0: Gibt die letzten Nachrichten des Benutzers zurück. Die letzte Nachricht dieser Ausgabe kann verwendet werden. Empfohlen für den ersten Programmstart um die ID der letzten Nachricht zu erhalten. conference_id=0 und message_id!=0: Gibt die letzten Nachrichten ab einer bestimmten Nachrichten-ID zurück. Dieser Fall kann für die Abfrage der Benachrichtigungen verwendet werden. conference_id!=0 und message_id=0: Gibt die letzten Nachrichten einer bestimmten Konferenz zurück. Kann für den Fall verwendet werden, wenn ein Benutzer zum ersten Mal eine Konferenz öffnet. Falls read=true: Konferenz wird als gelesen markiert (read=true, read_count=0, read_mid=letzte Nachrichten-ID). conference_id!=0 und message_id!=0: Gibt die letzten Nachrichten einer bestimmten Konferenz bis zu einer bestimmten Nachricht zurück. Kann verwendet werden, falls ein Benutzer beispielsweise weitere Nachrichten einer Konferenz anzeigen möchte. Falls read=true: Konferenz wird als gelesen markiert (read=true, read_count=0, read_mid=letzte Nachrichten-ID). Das Array-Element enthalten die folgenden Spalten: Die Nachrichten-ID.

conference_id: u64

Die Konferenz-ID.

user_id: u64

Die User-ID des Verfassers.

username: String

Der Username des Verfassers.

message: String

Die Nachricht.

action: String

Falls es sich um ein Befehl handelt, ist hier der Befehl angegeben.

timestamp: i64

Zeitstempel des Absendezeitpunktes.

device: String

Das Gerät/die Anwendung aus der eine Nachricht gesendet wurde.