Struct proxer::messenger::Messenger

pub struct Messenger<'messenger> { /* fields omitted */ }

Der Proxer Messenger ist ein Nachrichtensystem, welches es ermöglichen soll, über verschiedene Plattformen hinweg, mit anderen Proxer-Nutzern Nachrichten auszutauschen. Die aktuelle Schnittstelle basiert auf polling. Es ist eine zusätzliche Schnittstelle, die auf push-Nachrichten und Sockets basiert, geplant.

Methods

impl<'messenger> Messenger<'messenger>

fn get_constants(&self) -> Result<Vec<Constants>>

Liefert Messengerkonstanten. Bitte höchstens beim erstmaligen Start einer Anwendung durchführen. Diese Werte werden sich höchstens alle paar Monate mal ändern.

fn get_conferences(&self,
p_type: Option<ConferenceOption>,
p_page: Option<u64>)
-> Result<Vec<Conference>>

Liefert eine Liste der aktuellen Konferenzen.

Arguments

p_type- Gibt Konferenzen an, die eine bestimmte Markierung oder einen bestimmten Typ haben.

p_page- Die Seite der Konferenzen. Default ist 0.

fn get_conference_info(&self, p_conference_id: u64) -> Result<ConferenceInfo>

Informationen zu einer bestimmten Konferenz.

Arguments

p_conference_id - Die ID der Konferenz.

fn get_user_info(&self, p_user_id: u64) -> Result<UserInfo>

Informationen zu einem bestimmten Benutzer.

Arguments

p_user_id - Die ID des betroffenen Benutzers.

fn get_messages(&self,
p_conference_id: Option<u64>,
p_message: Option<u64>,
p_read: Option<bool>)
-> Result<Vec<Messages>>

Gibt die letzten Nachrichten einer Konferenz/eines Benutzers zurück.

Arguments

p_conference_id - Die Konferenz-ID.

p_message_id - Die Nachrichten-ID.

p_read - Ob eine Konferenz als gelesen markiert werden soll. Mögliche String-Werte: "true" (default) oder "false".

fn new_conference(&self, p_text: String, p_username: String) -> Result<u64>

Erstellt eine neue Unterhaltung, oder fügt eine Nachricht einer vorhandenen Unterhaltung hinzu (zwei Benutzer).

Arguments

p_text Eine Eingabenachricht. Beim Erstellen von Konferenzen werden Befehlseingaben ignoriert.

p_username Der Benutzername eines Proxer-Nutzers, an die eine Nachricht gesendet werden soll.

fn new_conferencegroup(&self,
p_users: String,
p_tropic: String,
p_text: Option<String>)
-> Result<u64>

Erstellt eine neue Gruppen-Unterhaltung.

Arguments

p_users Ein Array aus Benutzernamen. Benutzer, die einer Konferenz hinzugefügt werden sollen. Eingeschränkt durch die Konstante userLimit.

p_topic Das Thema/der Name der Konferenz. Eingeschränkt durch die Konstante topicCount.

p_text Die erste Nachricht der neu erstellten Konferenz. Eingeschränkt durch die Konstante textCount. Beim Erstellen von Konferenzen werden Befehlseingaben ignoriert.

fn report(&self, p_text: String, p_conference_id: u64) -> Result<()>

Funktionalität zum Melden einer Konferenz bei der Proxer-Administration.

Arguments

p_text - Ein kurzer Grund, weshalb die Konferenz gemeldet wird.

p_conference_id ID der Konferenz, die gemeldet werden soll.

fn set_message(&self, p_conference_id: u64, p_text: String) -> Result<String>

Sendet eine Nachricht an eine bestimmte Konferenz.

Arguments

p_conference_id - Die Konferenz-ID.

p_text - Eine Eingabenachricht. Die Nachricht ist eingeschränkt durch die Konstante textCount. Es gibt zwei Arten von Nachrichten: Normale Nachricht: Eine vom Benutzer abgesendete Nachricht. Eine Befehl: Ein Befehl fängt mit einem Schrägstrich an. Ein Beispiel für ein Befehl ist wie folgt: /addUser ProxerBot. Dieser Befehl würde den Benutzer ProxerBot zu der aktuellen Konferenz hinzufügen.

fn set_read(&self, p_conference_id: u64) -> Result<()>

Markiert eine Konferenz als gelesen (read=true, read_count=0). Beachte bitte: Funktion Get Messages.

Arguments

conference_id - Die Konferenz-ID.

fn set_unread(&self, p_conference_id: u64) -> Result<()>

Markiert eine Konferenz als ungelesen. Beachte bitte: Funktion Get Messages.

Arguments

conference_id - Die Konferenz-ID.

fn set_block(&self, p_conference_id: u64) -> Result<()>

Markiert eine Konferenz als blockiert.

Arguments

conference_id - Die Konferenz-ID.

fn set_unblock(&self, p_conference_id: u64) -> Result<()>

Hebt eine Blockierung einer Konferenz auf.

Arguments

conference_id - Die Konferenz-ID.

fn set_favour(&self, p_conference_id: u64) -> Result<()>

Markiert eine Konferenz als favorit.

Arguments

conference_id - Die Konferenz-ID.

fn set_unfavour(&self, p_conference_id: u64) -> Result<()>

Hebt eine Favorisierung einer Konferenz auf.

Arguments

conference_id - Die Konferenz-ID.