Module proxer::models

Structs

EmptyResponse

Response

Enums

ConferenceOption
Beschreibt die Konferenz genauer.

Firma
Gibt an durch welche Firma der Anime, Managa oder etc. veröffenlicht wurde.

Kategorie
Die Kategorie ist eine simple Unterscheidung zwischen Anime und Manga.

Language
Gibt die Sprache und die Übermittlung der Sprache an. (Über Untertitel oder Synchronisierung)

LengthLimit
Ob der Parameter "length" als Minimal- oder Maximalwert verwendet werden soll.

Medium
Gibt an um was für eine Medienquelle es sich handelt.

SearchSort
Nach welchem Parameter bei der Suche sotiert werden soll.

Sort
Nach was die angefragte List sortiert werden soll.

SubType
(Betrifft nur Tags vom Typ entry_tag, alle anderen Typen haben immer Kategorie "Sonstige"). Default: Alle. Erlaubte Werte:

TranslationStatus
Gibt den Stand der Übersetzung an.

WatchType
Beschreibt ob das Medium noch geschaut wird, Favoriesiert ist oder Abgeschlossen ist.