Struct proxer::notification::Notification

pub struct Notification<'notification> { /* fields omitted */ }

Diese Klasse beinhaltet alle Schnittstellen, die mit Daten zu tun haben, die normalerweise auf Proxer oben rechts bei den Notifications zu sehen sind, insbesondere News und Benachrichtigungen.

Methods

impl<'notification> Notification<'notification>

fn get_count(&self) -> Result<String>

Diese Funktion gibt die Anzahl an verschiedenen Notification-Kategorien zurück (kleine rote Zahlen auf Proxer). Die Ausgabedaten befinden sich in einem Array, wobei folgende Indices folgende Anzahlen enthalten:

0 = Error (entspricht error-Wert des jsons)

1 = Altes PN-System (deprecated)

2 = Neues PN-System

3 = Freundschaftsanfragen

4 = News

5 = Benachrichtigungen

fn get_news_per_api(&self,
p_page: Option<u64>,
p_limit: Option<u64>)
-> Result<Vec<News>>

Gibt die neuesten News aus. Der Bildlink einer News setzt sich zusammen aus: cdn.proxer.me/news/[News-ID]_[Image-ID].png Für Tumbnail: cdn.proxer.me/news/th/[News-ID]_[Image-ID].png Link zum Forumspost der News: proxer.me/forum/[catid]/[mid]

Arguments

p - Die zu ladende Seite, beginnend ab 0 (Auf Seite 0 befinden sich die neuesten News, nach hinten werden die News älter). Wenn nicht gegeben, so wird die erste Seite geladen.

limit - Die Anzahl der zu ladenden News pro Seite. Default 15.

fn delete_notification(&self, p_nid: Option<u64>) -> Result<()>

Löscht eine gegebene Notification

Arguments

p_nid - Die ID der zu löschenden Notification. Wenn weggelassen oder 0, so werden alle als gelesen markierten Benachrichtigungen gelöscht.