Struct proxer::OldNotification

pub struct OldNotification {
    pub nid: u64,
    pub time: i64,
    pub description: String,
    pub image_id: u64,
    pub image_style: String,
    pub subject: String,
    pub hits: u64,
    pub thread: u64,
    pub uid: u64,
    pub uname: String,
    pub posts: u64,
    pub catid: u64,
    pub catname: String,
}

Beinhaltet eine einzelene News bzw. Notification (Benutzt die alte api Schnittstelle?).

Eingabe

http://proxer.me/notifications?format=json&s=news&p=1 Der Parameter p gibt die aktuelle Seite an. Hinweis: Aktuell wird der Abruf der Seitennummer ignoriert. Author: souryo(proxer-rs)

Arguments

nid - Die Nachrichten ID.

time - Ein Unix Timestamp der angibt, wann die Nachricht veröffentlicht wurde.

mid - Veraltet. Bitte stattdessen "thread" verwenden.

pid - Veraltet.

description - Eine Kurzbeschreibung zu der News.

image_id - Die Bild-ID der News.

image_style - Hier sind CSS Style-Angaben zum Bild gespeichert.

subject - Die Überschrift der News.

hits - Anzahl der News-Aufrufe.

thread - Die Thread-ID. Ist äquivalent zu mid.

uid - Die Benutzer-ID des Autors.

uname - Der Benutzername des Autors.

posts - Anzahl der Antworten/Kommentare auf die News.

catid - Die Kategorie-ID der Kategorie, in der sich die News befindet.

catname - Der Name der Kategorie, in der sich die News befindet.

Fields
nid: u64

News id

time: i64

Time als Unix-Timestamp

description: String

Die Beschreibung

image_id: u64

Die Bild id

image_style: String

Bildstyles

subject: String hits: u64

Aufrufe der Notification

thread: u64

thread id

uid: u64 uname: String posts: u64 catid: u64 catname: String
Methods

impl OldNotification

fn get_image(&self) -> String

Der Link zum Bild ist folgendermaßen aufgebaut: http://cdn.proxer.me/news/{nid}_{image_id}.png (Souryo: Sollte eigentlich nicht von belangen sein, dennoch hab ich es der Vollständigkeits halber mit dazu genommen) Beachte, dass hier nur Thumbnails ausgegeben werden. Falls Zugriff auf die Originalbilder nötig ist, kann genesis(http://proxer.me/wiki/Benutzer:Genesis) hierzu angeschrieben werden.

fn get_news_link(&self) -> String

Der Link zur News ist folgendermaßen aufgebaut: http://proxer.me/forum/{catid}/{thread} Alternativ kann an den Link der Anker #top angegeben werden. (Souryo: Sollte eigentlich nicht von belangen sein, dennoch hab ich es der Vollständigkeits halber mit dazu genommen)