Struct proxer::ucp::Ucp

pub struct Ucp<'ucp> { /* fields omitted */ }

Diese Klasse dient der Verwaltung sämtlicher Daten, die normalerweise über das UCP Abrufbar/Veränderbar sind. Logischerweise erfordern alle diese Funktionen, dass der User eingeloggt ist.

Methods

impl<'ucp> Ucp<'ucp>

fn get_list(&self,
p_kat: Option<Kategorie>,
p_page: Option<u64>,
p_limit: Option<u64>,
p_search: Option<String>,
p_search_start: Option<String>,
p_sort: Option<Sort>)
-> Result<Vec<List>>

Diese Funktion liefert die Liste aller Animes/Mangas, zu denen der User einen Eintrag im UCP hat.

Arguments

p_kat - Die Kategorie, die geladen werden soll. Mögliche Werte: anime, manga. Default: anime.

p_page - Dieser Parameter gibt an, welche Seite geladen werden soll. Default Wert 0. Start bei 0.

p_limt - Dieser Parameter gibt an, wie viele Einträge eine Seite enthalten soll. Default Wert 100.

p_search - Durch die Angabe dieses Parameters werden nur Entrys angezeigt, die den angegeben Wert als Substring ihres Namens haben. Dabei ist die Position im Namen egal.

p_search_start - Durch die Angabe dieses Parameters werden nur Entrys angezeigt, die den angegeben Wert als Substring zu Beginn ihres Namens haben.

p_sort - Dieser Parameter gibt an, wie die Liste sortiert werden soll, erlaubte Eingaben (Fehlerhafte Eingaben werden auf den Default-Wert gezwungen)

fn get_listsum(&self, p_kat: Option<Kategorie>) -> Result<String>

Diese Funktion liefert die Summe der Episoden/Kapitel, die der User bisher gesehen hat.

Arguments

p_kat - Die Kategorie, die geladen werden soll. Mögliche Werte: anime, manga. Default: anime.

fn get_topten(&self) -> Result<String>

Diese Funktion liefert die Top-Ten des Users. (Anime und Manga)

fn get_history(&self,
p_limit: Option<u64>,
p_page: Option<u64>)
-> Result<Vec<History>>

Diese Funktion liefert die Chronik des Users.

Arguments

p_limit - Dieser Parameter gibt an, wie viele Einträge eine Seite der Chronik haben soll. Default Wert 50.

p_page - Dieser Parameter gibt an, welche Seite der Chronik geladen werden soll. Start bei 0, Default Wert 0.

fn get_votes(&self) -> Result<Vec<Vote>>

Diese Funktion liefert die Kommentarvotes des Users.

fn get_reminder(&self,
p_kat: Option<Kategorie>,
p_page: Option<u64>,
p_limit: Option<u64>)
-> Result<Vec<Reminder>>

Diese Funktion liefert eine Liste aller Lesezeichen des Users.

Arguments

p_kat - Dieser Parameter gibt an, welche Kategorie geladen werden soll. Wenn weggelassen werden beide Kategorien geladen. Erlaubte Werte: "anime","manga"

p_page - Die zu ladende Seite, Start bei 0. Default 0

p_limit - Die zu ladenden Einträge pro Seite. Default 100

fn delete_reminder(&self, p_id: u64) -> Result<()>

Diese Funktion löscht ein Lesezeichen.

Arguments

p_id - Die ID des zu löschenden Lesezeichens (erhältlich über die "Reminder" Funktion)

fn delte_favorite(&self, p_id: u64) -> Result<()>

Diese Funktion löscht einen Eintrag der Top-ten.

Arguments

p_id - Die ID des zu löschenden Eintrags (erhältlich über die "Favorite" Funktion)

fn delte_vote(&self, p_id: u64) -> Result<()>

Diese Funktion löscht einen Kommentarvote.

Arguments

p_id - Die ID des zu löschenden Eintrags (erhältlich über die "Vote" Funktion)

fn set_commentstate(&self, p_id: u64, p_value: u64) -> Result<()>

Diese Funktion setzt die Zahl der bereits gesehenen/gelesenen Folgen/Kapitel für einen Anime/Manga aus der Liste des Users. Wird dieser Wert auf oder über die Zahl der vorhandenen Folgen/Kapitel gesetzt, so wird zudem der Status des Animes/Mangas auf "Abgeschlossen" gesetzt.

Arguments

p_id - Die ID des zu bearbeitenden Eintrags (erhältlich über die "List" Funktion)

p_value - Der zu setzende Wert

fn set_reminder(&self, p_id: u64, p_value: u64) -> Result<()>

Diese Funktion setzt eine Episode auf die Watch/Readlist eines eingeloggten Users.

Arguments

p_id - Die id des Entrys

p_episode - Die Episodennummer, auf die das Lesezeichen gesetzt werden soll.

p_language - Die zu ladende Sprache. (Für Animes: gersub,gerdub,engsub,engdub; Für Mangas: de,en)

p_kat - Die Kategorie des Entrys (manga oder anime)