Struct proxer::user::User

pub struct User<'user> {
    pub uid: u64,
    pub avatar: String,
    pub token: String,
    // some fields omitted
}

Diese Klasse beinhaltet die grundlegenden Schnittstellen zur Verwaltung von Userdaten, vornehmlich Login und Logout, aber auch die Möglichkeit, einen neuen User zu registrieren sowie die öffentlichen Daten eines jeden Users per ID oder Username abzufragen.

Fields
uid: u64

Die ID des eingeloggten Users.

avatar: String

Der Avatar des eingeloggten Users.

token: String

Ein Login-Token. Möchte man das gewöhnliche Cookie-basierte Login System nicht verwenden, so kann man stattdessen bei jeder Anfrage die einen Login erfordert dieses Token senden.

Methods

impl<'user> User<'user>

fn login(p_proxer: &Proxer, p_username: &str, p_password: &str) -> Result<Login>

Mit dieser Schnittstelle kann ein User mithilfe eines Passwortes und eines Usernamen eingeloggt werden

Arguments

p_username - Der Benutzername des zu einloggenden Benutzers.

p_password - Das Passwort des zu einloggenden Benutzers

p_api_key - API_KEY.

fn logout(self) -> Result<()>

Mit dieser Schnittstelle kann ein User ausgeloggt werden

fn get_userinfo(&self) -> Result<UserInfo>

Mit dieser Schnittstelle können die öffentlichen Daten jedes Users per ID oder Username abgefragt werden. Sind ID und Username gegeben, so wird ausschließlich die ID verwendet. Ist weder ID noch Username gegeben, so werden die Daten des eingeloggten Users abgerufen. Ist auch dies nicht gegeben, so wird eine Fehlermeldung ausgegeben.

Arguments

&self - User-ID, deren Daten abgefragt werden sollen

fn get_topten(&self,
p_uid: u64,
p_kat: Option<Kategorie>)
-> Result<Vec<TopTen>>

Mit dieser Schnittstelle können die Topten jedes Users per ID oder Username abgefragt werden. Sind ID und Username gegeben, so wird ausschließlich die ID verwendet. Ist weder ID noch Username gegeben, so wird eine Fehlermeldung ausgegeben. Diese Schnittstelle respektiert die Sichtbarkeitseinstellungen der User. Hat ein User etwa die Sichtbarkeit nur "Für Freunde", so wird geprüft ob der momentan eingeloggte Nutzer mit dem User Befreundet ist. Es ist daher ratsam, vor der Verwendung dieser Schnittstelle einen User einzuloggen (oder ein token zu verwenden).

Arguments

p_uid - User-ID, deren Daten abgefragt werden sollen

p_kat - Die Kategorie, die geladen werden soll. Mögliche Werte: anime, manga. Default: anime.

fn get_list(&self,
p_uid: u64,
p_kat: Option<Kategorie>,
p_page: Option<u64>,
p_limit: Option<u64>,
p_search: Option<String>,
p_search_start: Option<String>,
p_sort: Option<Sort>)
-> Result<Vec<List>>

Mit dieser Schnittstelle können die Entrylisten jedes Users per ID oder Username abgefragt werden. Sind ID und Username gegeben, so wird ausschließlich die ID verwendet. Ist weder ID noch Username gegeben, so wird eine Fehlermeldung ausgegeben. Diese Schnittstelle respektiert die Sichtbarkeitseinstellungen der User. Hat ein User etwa die Sichtbarkeit nur "Für Freunde", so wird geprüft ob der momentan eingeloggte Nutzer mit dem User Befreundet ist. Es ist daher ratsam, vor der Verwendung dieser Schnittstelle einen User einzuloggen (oder ein token zu verwenden).

Arguments

p_uid - User-ID, deren Daten abgefragt werden sollen

p_kat - Die Kategorie, die geladen werden soll. Mögliche Werte: anime, manga. Default: anime.

p_page - Dieser Parameter gibt an, welche Seite geladen werden soll. Default Wert 0. Start bei 0.

p_limt - Dieser Parameter gibt an, wie viele Einträge eine Seite enthalten soll. Default Wert 100.

p_search - Durch die Angabe dieses Parameters werden nur Entrys angezeigt, die den angegeben Wert als Substring ihres Namens haben. Dabei ist die Position im Namen egal.

p_search_start - Durch die Angabe dieses Parameters werden nur Entrys angezeigt, die den angegeben Wert als Substring zu Beginn ihres Namens haben.

p_sort - Dieser Parameter gibt an, wie die Liste sortiert werden soll, erlaubte Eingaben (Fehlerhafte Eingaben werden auf den Default-Wert gezwungen)

fn get_latestcomments(&self,
p_kat: Option<Kategorie>,
p_page: Option<u64>,
p_limit: Option<u64>,
p_length: Option<u64>)
-> Result<Vec<LatestComment>>

Mit dieser Schnittstelle können die letzten Kommentare jedes Users per ID oder Username abgefragt werden. Sind ID und Username gegeben, so wird ausschließlich die ID verwendet. Ist weder ID noch Username gegeben, so wird eine Fehlermeldung ausgegeben. Diese Schnittstelle respektiert die Sichtbarkeitseinstellungen der User. Hat ein User etwa die Sichtbarkeit nur "Für Freunde", so wird geprüft ob der momentan eingeloggte Nutzer mit dem User Befreundet ist. Es ist daher ratsam, vor der Verwendung dieser Schnittstelle einen User einzuloggen (oder ein token zu verwenden).

&self - User-ID, deren Daten abgefragt werden solle.

p_kat - Die Kategorie, die geladen werden soll. Mögliche Werte: anime, manga. Default: anime.

p_page - Dieser Parameter gibt an, welche Seite geladen werden soll. Default Wert 0. Start bei 0.

p_limit - Dieser Parameter gibt an, wie viele Einträge eine Seite enthalten soll. Default Wert 25.

p_length - Dieser Parameter gibt die minimale Anzahl an Zeichen an, ab der ein Kommentar angezeigt werden soll. Default 300.